10598: Move refresh_schema logic into override itself
